Rename package from standardfunctionhandler to library
