+++ trunk/source/CACSL2BoogieTranslator/src/de/uni_freiburg/informatik/ultimate/cdt/translation/implementation/base/CHandler.java
@@ -180,23 +180,23 @@ import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.c
 import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.chandler.TypeSizeAndOffsetComputer;
 import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.chandler.TypeSizes;
 import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.expressiontranslation.ExpressionTranslation;
-import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.standardfunctions.AssertLibraryModel;
-import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.standardfunctions.AtomicLibraryModel;
-import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.standardfunctions.FloatLibraryModel;
-import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.standardfunctions.FunctionModelHelper;
-import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.standardfunctions.GccBuiltinLibraryModel;
-import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.standardfunctions.ILibraryModel;
-import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.standardfunctions.LibraryModelHandler;
-import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.standardfunctions.LinuxLibraryModel;
-import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.standardfunctions.PthreadLibraryModel;
-import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.standardfunctions.SetjmpLibraryModel;
-import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.standardfunctions.SocketLibraryModel;
-import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.standardfunctions.StdioLibraryModel;
-import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.standardfunctions.StdlibLibraryModel;
-import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.standardfunctions.StringLibraryModel;
-import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.standardfunctions.SvcompLibraryModel;
-import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.standardfunctions.TimeLibraryModel;
-import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.standardfunctions.VariadicLibraryModel;
+import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.library.AssertLibraryModel;
+import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.library.AtomicLibraryModel;
+import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.library.FloatLibraryModel;
+import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.library.FunctionModelHelper;
+import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.library.GccBuiltinLibraryModel;
+import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.library.ILibraryModel;
+import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.library.LibraryModelHandler;
+import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.library.LinuxLibraryModel;
+import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.library.PthreadLibraryModel;
+import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.library.SetjmpLibraryModel;
+import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.library.SocketLibraryModel;
+import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.library.StdioLibraryModel;
+import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.library.StdlibLibraryModel;
+import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.library.StringLibraryModel;
+import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.library.SvcompLibraryModel;
+import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.library.TimeLibraryModel;
+import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.library.VariadicLibraryModel;
 import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.container.AuxVarInfo;
 import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.container.AuxVarInfoBuilder;
 import de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.container.SymbolTableValue;
