+++ trunk/source/CACSL2BoogieTranslator/src/de/uni_freiburg/informatik/ultimate/cdt/translation/implementation/base/library/FloatLibraryModel.java
@@ -1,4 +1,4 @@
-package de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.standardfunctions;
+package de.uni_freiburg.informatik.ultimate.cdt.translation.implementation.base.library;
 
 import java.util.ArrayList;
 import java.util.Arrays;
